requirements.txt
pytest==7.3.1
google-cloud-firestore==2.11.1
python-dotenv==1.0.0
